Update README.md
How do the rich live? We see them on television playing polo and living the high 
life. What are they willing to endure to live that life? Who are these people? Why do 
they skip the drudgery and pain of financial struggle that most seem to endure? 
They pay less in taxes and earn more with seemly no effort. What are their secrets?